Temp: Tag RDS instances
Our new RDS broker keeps track of which instances were deployed by it by
using AWS tags. To bootstrap this functionality, we need to tag instances
that are already deployed. Use this script and concourse task to do so.
This can be removed once we have deployed new broker into production.