Temp: Update the RDS broker master passwords
This script will be added to the pipelines to update all the RDS broker
master passwords of all the RDS instances with the new password.

We use a shell script which queries all the RDS instances in the
environment and updates the master password. The script is idempotent.

Once https://github.com/alphagov/paas-cf/pull/314 has been widely
deployed, this commit can be reverted.

You can test if this change work by binding any existing DB to a new
app, e.g:

  cf bind-service someapp healthcheck-db

Note: RDS takes some time to update the master password.