Change UAA logging level from DEBUG to INFO
Reduce UAA's logging level from the spec default of `DEBUG` to `INFO`. This
will make the logs easier to read and identify errors, as well as reducing
the amount of logs that Logsearch has to parse and store.

There doesn't seem to be much order to the `uaa` property keys. So I've
added the new property to what I think is a vaguely sensible place.

We did the same for CC in 62c0882. UAA is the last component to still log at
debug level, according to the following Logsearch query:

    debug AND NOT syslog_sd_params.bosh_template:uaa AND NOT syslog_program:vcap.uaa_ctl*

I've raised an upstream PR to change the default in future:

- cloudfoundry/uaa-release#20