Add base test to check availability during deploy
Add a base test to monitor the `healthcheck app` to be sure that there is
not downtime during deployment.

We implement the test using vegeta[1], which is the same library than
the one used for performance test. This library is able to execute the
queries in a separated worker tracking all the responses. Vegeta provides
an API to query different metrics which we can check in the test.

In this case we run vegeta with one worker, with a rate of 10 reqs/sec.
It will run forever until we call the corresponding stop function
at the end of the test.

We use Eventually[1] from gomega to check a criteria which will
determine when the test should be finish. This criteria is implemented
in variable pointing to a function `stopAttackCriteria()`.

In this commit we only provide a dummy implementation, which always
returns true. We will implement a proper check in next commits.
We also add logic to stop the test if the rate of error after
15 seconds is > 50% in case this is the first deployment and healthcheck
is not available at all.

In order to identify the query in `healthcheck app` logs, we add the
parameter `test-availability=<pipeline-trigger-version>`, being the
`pipeline-trigger-version` the value of resource `pipeline-trigger`
which identifies the pipeline execution uniquely.

We get all the configuration using environment variables. The reason
is that we will not have all the information required to build
the `config.json` used in the CF test helpers and because we need
some additional configuration.

We print a vegeta report to stdout at the end of the test for information.

[1] https://github.com/tsenart/vegeta/
[1] https://onsi.github.io/gomega/#making-asynchronous-assertions