Add pre-deploy job to wait for availability tests
The availability test takes some time to start monitoring healthcheck
(concourse needs to start the container, the go code must compile, etc).

Because that we add a pre-deploy job that will use `cf logs` to check if
the test is hitting the healthcheck app. We use pipeline-trigger version
to check that the request are related to the current deployment.