Add cflinuxfs2-rootfs release to manifest.
This was split out from the diego release in version 0.1469.0 [1]. As a
result of this, some of the property names have also changed.

Now that this is in a separate release, the diego `rep` job no longer
contains a default for the `preloaded_routefses` property. This needs to
be set, and the diego manifest-generation scripts set this to the
relevant value for the new cflinuxfs2 release[2]. We therefore use this
property in our manifest.

[1]https://github.com/cloudfoundry-incubator/diego-release/releases/tag/v0.1469.0
[2]https://github.com/cloudfoundry-incubator/diego-release/blob/v0.1472.0/manifest-generation/diego.yml#L592