Update graphite-nozzle to use Go 1.6
The current version of the Go buildpack no longer supports Go 1.4 which
the graphite-nozzle app was pinned to. This updates the app to a version
which uses Go 1.6.

This also pins the buildpack in the application manifest to avoid having
to use buildpack detection, which will speed up the deployment slightly.