Fix delete-deployment when testers org doesn't exist.
At present the remove-healthcheck-db task in the destroy-cloudfoundry
and autodelete-cloudfoundry pipelines will error if the CF api is
available, but the testers org doesn't exist. This can happen if the
cf-deployment succeeds, but one of the post-deploy tasks fails for some
reason.

This updates the task to check for the exsitence of the testers org
before attempting to delete the healthcheck app and db. If the org isn't
present, the app and db can't exist, so it's safe to just skip the task.