Remove logsearch-shipper
Logsearch-shipper was added to all hosts during the deployment of
elasticsearch however it was not noticed that we had also configured
Metron to ship all syslog logs over RELP to the same logsearch cluster.
This resulted in duplicated logs.

Logsearch-shipper also contains a collectd install that is configured to
report all metrics over graphite to port 2003 on localhost, where a
netcat listener then writes them to disk as a logfile, which is then
picked up by nxlog and sent to the logsearch cluster as the `metric`
type.

https://github.com/logsearch/logsearch-shipper-boshrelease/blob/be4c7c4901fbe9dd150eceefbe9ece00115a6d4a/jobs/logsearch-shipper/templates/bin/metrics-collectd.graphite-mock

Given that we don't really want to have two collectd instances running
(one of which is just dumping useless data into our logsearch cluster)
and we don't want to have duplicated logs, we decided to remove
logsearch-shipper as metron handles all log shipping anyway and the
new collectd install will replace the useless one currently deployed.

We spent a substantial amount of time spelunking in kibana looking for
logs that were only being reported by logsearch-shipper and not metron
and only found one example, which is the syslog drain used by metron. If
it was logging to syslog then it would go into an infinite loop, so we
think this is reasonable.