A script to unbind paas-cf in the given job
This script reads pipeline as input and removes passed for paas-cf
resource in the specified job. It outputs modified pipeline.

Because 'plan' can contain steps directly, but also 'do' and
'aggregate', which in turn can contain further steps and again
'do' and 'aggregate', we have to traverse the data structure
recursively and remove 'passed' from all occurrences of
`- get: paas-cf` we encounter - as it is possible various job
could be retrieving the resource multiple times.