Add a way to unbind and run a job in create-bosh-cf
This adds:
  * README updates that describe new `run_job` target
  * MAKEFILE `run_job` target
  * run_job.sh script

This script unbinds (removes passed on paas-cf) for the specified job
in create-bosh-cloudfoundry pipeline and then triggers that job.

We retrieve already deployed pipeline from concourse because PSYCH has
some processing issues with fly {{variables}}, meaning we can't run this
script on the pipeline template with unexpanded variables. Retrieving
pipeline from concourse also preserves all other changes user might
have done.

We remove passed from job and then upload pipeline in two steps. The
reason is that retrieval, but also passed removal can fail. In that case
script exits and prevents possible upload of bogus/empty pipeline back
to concourse.

Using pipelines or named pipes to synchronize these 2 steps is not
possible, because pipelines would give empty input into fly set-pipeline
in case of any previous steps failing. This would then effectively
delete the pipeline (make it empty). Named pipes on the other hand are
blocking because of the fixed buffer size, which means that you need to
execute passed removal in background in order to not block. Unblocking
happens by running fly set-pipelines, which reads from the pipe. If
there is any problem with downloading of the pipeline or passed
removal, set-pipelines is running already and the situation is the same
as if you were using standard pipelines. Meaning set-pipeline can
upload broken/empty pipeline. Thus we render to a temp file and then
upload from there.