Set gorouter drain_wait to 15s
With this setting the gorouter will keep serving new connections for
15 seconds after getting the drain signal, but respond 503 error when
the HTTPMonitor User-Agent header is used. This header is inserted by
using a custom frontent in haproxy

The ELB healthchecks run every 5 seconds. After 2 healthcheck failures,
the backend will be declared unhealthy. It may then take up to 10s for
the ELB to deregister a backend, hence the 15s drain_wait to cover this
period

This way we ensure that we don't drop any incoming connection before the
ELB marks the gorouter down.