Fix capture of bosh-init state on error.
Because the updated state file was only copied into the output directory
after the bosh-init run, if bosh-init errored for any reason, this would
not happen, and the ensure block had no file available to upload.

This refactors the task to use the output dir as the working dir for
bosh-init meaning that the updated state file is always available to the
ensure block.