Add property ha_proxy.enable_healthcheck_frontend
This adds an HAProxy frontend listening on port 82 that we can use
for healthchecking gorouter (it adds a custom header recognised by
gorouter).

This property is implemented in the release 0.0.3

The ELB healthcheck should later be changed to `HTTP:82/`. This will be
done in a separate deploy AFTER the new frontend has been added.