Move availability test helpers into GOPATH.
This avoids having to refer to them as an external dependency tracked in
the Godeps.json. This in turn could lead to odd versioning issues where
the helpers code being used wasn't the version alongside the tests in
the repo, and instead was a version from a previous commit that's
referenced in the Godeps.json.

Moving these underneath `availability` in the GOPATH means that godep
doesn't pin them in the Godeps.json. This means that the tests will
always use the version of these directly from the repo.