Create results channel before consuming from it.
This reorders the code so that the resultChannel is created before we
start the goroutine to read from it. Otherwise there's a race with the
read from the channel, and the assignment to the channel variable, which
could lead to problems. It's also easier to see what's going on if we
create the channel before consuming results from it.

This also removes the unnecessary `metrics.Close()` from the results
consuming goroutine. This is called in the main function before the
aggregate values are used. The name of this methods is a little
misleading - this actually is just used to calculate all the derived
summary metrics.  It doesn't actually close anything.