Use ExpectWithOffset in test helpers
This has largely the same behaviour as `Expect`. It only differs when it
comes to reporting failures. The offset is the number of callers to
unwind when repoiting the line number the failure happened on. This
makes it easier to see which call to this function triggered the error.

See http://onsi.github.io/gomega/#making-assertions-in-helper-functions
for more details.