Set system_domain_organization to 'admin'
This was set to 'GDS' in 11318b0 for no real reason that I can remember
(I was working on that story). Having it set to 'GDS' could lead to
confusion as there are likely to be several services deployed to the
PaaS from teams within GDS.

This sets it back to 'admin' which is the cloudfoundry default. The
admin org is also where we deploy some system components like the
graphite nozzle, so seems like a good org to own this.