Temporary pipeline task to fixup system domain org
Changing the `system_domain_organization` in the manifest doesn't result
in this being applied when deployed to an existing deployment. This
therefore adds a task to remove the old org (and thereby delete the
domain), and add the system domain to the admin org.

This should be reverted once it's run on production