Allow cloudcontroller to talk to service brokers
This creates a new security group `cloud_controller` for the
cloudcontroller components. A rule is then added to the
`service_brokers` security group to allow cloudcontroller to connect.

This is a break from the existing pattern where there's a client
security group for each service. Instead this moves towards having a
security group for each component, and having each service's security
group define rules specifying which components are allowed to connect.
This makes it clearer which components are talking to each service. It
also means we're less likely yo run into the AWS limit of 5 security
groups per network interface[1].

[1]http://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_vpc