Remove domain property in favour of system_domain
Previously these 2 properties both existed, and were set to the same
value. This has led to some confusion as there were 3 properties -
`domain`, `system_domain` and `apps_domains`. From this it wasn't clear
what the `domain` property was for.  Removing it makes it much clearer.

This change has also been made in the upstream manifests.