Remove collector (stats)
Remove collector VM (stats/0). We were not even configuring it, so this
VM is simply just sitting there and doing nothing.

In the logs it looks like it is sending metrics, but there are no
endpoints configured for historian, so it doesn't send it anywhere.

Moreover, colllector is being deprecated in general. For more details,
see for example
https://lists.cloudfoundry.org/archives/list/cf-dev@lists.cloudfoundry.org/message/MBAR3MQK6B6B23S7SEGOUIIAW6FDQNUB/