Alert on temp. admin removal failure
We have noticed that we are having a lot of temporary admin user removal
failures. We would like to receive notification when this happens, as
this indicates that there are possible issues with UAA and also leaves
an admin user behind. This can impact UAA (as we are leaking an user)
and security (as there would be many admin users that are not supposed
to exist).

We have decided to use new notification script, because current
similarities with smoketest failure notification will get even smaller
when we add further smoketest failure info back into that notification.