Revert "Remove etcd_metrics_server"
This reverts commit b3092a1da86a3236f93f67a51ba3ed74506bc18f.

As it turns out, ETCD metrics server was also sending metrics to doppler,
meaning they would make it to our graphite nozzle which was adding them
to graphite. Add this component back so that we have ETCD metrics again.

Unfortunately no etcd_metrics_server documentation mentioned this
feature. It was added in this commit:
https://github.com/cloudfoundry-incubator/etcd-metrics-server/commit/0668d4fcbb66105d1c13f6d72ac130792b52cc26