Ensure bbs is upgraded before cells
When we upgraded from diego v0.1472.0 to diego v0.1476.0 previously deployed
apps started returning 404 errors.

We found that:

- the cf CLI showed application instances in the "starting" state
- the corresponding containers and application processes were actually running
  on the cells
- veritas dump-state [1] showed corresponding LRP instances in the CLAIMED
  state
- the rep was logging an error indicating that a timeout occurred [2]. The
  message was timestamped 0ms after the event started.

We talked to the diego team and they said a migration [3] was written to
migrate all timeouts from seconds to milliseconds.

Previously this migration was being applied after the cells were upgraded, so
container instances were being started with no timeout_ms settings. In go this
means that it defaults to 0ms.

Note: This also ensures that the database VMs are updated before the
cells, even though we're not using them at present. This is to ensure
that this problem doesn't regress if/when we move away from the
colocated VMs.

[1] https://github.com/pivotal-cf-experimental/veritas
[2] The full message is a bit too long for a commit message, but the name is
rep.executing-container-operation.ordinary-lrp-processor.process-reserved-container.run-container.containerstore-run.node-run.monitor-run.timeout-step.timed-out
[3] https://github.com/cloudfoundry/bbs/blob/6c88d108713485116c1a2425c140b0104eab61ea/db/migrations/1451635200_timeouts_to_milliseconds.go