Order diego jobs to match upstream manifest.
These jobs were initially added in 80ab447eb, and look to have been
added in alphabetical order. This differs from the upstream
manifests[1] which have a specific order.

We got caught out by this by an issue when the cells were upgraded
before BSS (running on the colocated VMs). Maintaining these in the same
order as upstream will help avoid this sort of issue in future.

Unlike the upstream manifests, this puts the colocated definition first
instead of last. This is because we're currently running a mixed
deployment where not everything is on the colocated VMs (we've split the
cells out). Putting this first ensures that BBS is updated before the
cells.

[1]https://github.com/cloudfoundry/diego-release/blob/v0.1476.0/manifest-generation/diego.yml#L122-L424