Add reminder to change password
See [Include Change Password Instructions in Welcome Mail][1] for
details. There is no point putting the actual command to change
password in the welcome email, because that won't work until user is
logged in, but this reminder makes clear that the password that was
included in the email can be changed by the tenant.

Also fix email copy.

[1]: https://www.pivotaltracker.com/story/show/125537351