Auto-refresh user-impact dashboard every 30s
So that we can see the most recent metrics, rather than the metrics at the
time the tab was last loaded or manually refreshed.

We could have chosen an auto-refresh value of 10s or more, because that's
the max resolution we collect metrics at. Alex and I decided that 10s was
probably too frequent for a large screen that we only look at occasionally
and it *could* generate some load on Graphite. 1m didn't seem frequent
enough. So we went for 30s, which is in the middle.

The value can be changed in your browser if you're viewing the dashboard
locally on your own laptop.

Our dashboard screen will need to be manually refreshed to pick up this
change.