Collectd: Add limit on metric buffer
Add a limit of 100k messages on the collectd in-memory metric buffer.

This buffer gets used when collectd is failing to send metrics to the
configured destinations. By default it is unlimited and can cause
collectd to consume excessive amounts of memory if the server it is
supposed to send metrics to is unavailable. This then leads to OOM
killer activation - but which can also kill some other processes on
the VM than collectd.

With 100k message limit, collectd will buffer up to 50k metrics. Then
it will start dropping proportional amount to remaining buffer space.
Meaning that if it is storing 75k metrics already, it will drop every
second metric. This way you lose some resolution, but you are able to
buffer metrics for longer time. More details described here:

https://collectd.org/documentation/manpages/collectd.conf.5.shtml
search for 'WriteQueueLimitHigh'