Set backup retention of RDSbroker to 7 days
We want all the tenant DB instances to have at a retention period
of 7 days. We decided this period to start working on the
related stories to backup and restore tenant data.

Later we might change this value or allow the tenants to customise it.

We add tests for the RDS broker pg plans to ensure that all the configured
plans have the same retention period of 7 days.

We also add a test to be sure that only the defined and tested
plans are included. We were not checking it and we could be offering
additional plans without testing.