Remove paas_demo and trade-tariff quota definitions
These are identical to the newly created 'small' and 'medium' quotas, so
we're going to update the relevant orgs to use the standard quotas.

Note: removing these from the manifest here won't cause the quotas to be
deleted from CF (the quota setting script only does create/update). This
will have to be done separately once the orgs have been updated to use
the standard quotas.