Use awscli version with get-caller-identity
We need to use the `aws sts get-caller-identity`, required to
retrieve the account ID to generate the ARN of the RDS instances
in the script that updates the RDS instances. This is the only command
that can work on deployer with the current policy.

This command is only available in versions > 1.10.18, and we install
an older version in the requirements.txt.

This file is currently used only for local development to install
all the dependencies.

We update the repository to the same version than the one in
the awscli container used in concourse[1]

[1] https://github.com/alphagov/paas-docker-cloudfoundry-tools/blob/master/awscli/Dockerfile#L3