Remove colocated from failure-testing pipeline
We're no longer running the colocated VMs, so this pipeline wouldn't
work with this in place.

We've decided not to add anything in its place at this time because at
present this is more of a proof of concept. At the point that we start
to run this regularly, we'll look at expanding it's coverage.