Set Kibana timezone to UTC
This script will change the default timezone used by Kibana to display
timestamps. It emits a proper setting directly into the `.kibana`
index stored in elasticsearch.

It will also create a `.kibana` index if one doesn't exist, with the default
settings normally used by Kibana: 1 shard with 1 replica only.

After the release of Kibana 5 we will refactor this script to use Kibana
API: https://github.com/elastic/kibana/pull/6849

Test that the script produces the expected exit codes and hits the expected endpoints.

We run the script under test in a separate process since that is closer to reality.
This requires us to mock http endpoints using a real http server.
We had to upgrade `mimic` and work around a few design issues as the tooling for doing this isn't very good.
We might want to reconsider how we do this sort of testing if we have to do a lot of it.

@combor & @benhyland