Remove unneeded org-specific quotas.
Now that we have some standard quotas in place, these quotas are no
longer needed.

The orgs that were previously using these quotas have been switched to
using the new standardised quotas. This was done by running commands
like `cf set-quota foo-org small` for all the orgs that were using these
quotas. The new standardised quotas were created to be identical to
these quotas, so there will be no impact to the orgs as these are
switched, the only visible change will be the name of the quota applied
to the org.

Once this has run everywhere, it can be reverted.