Keep only last 10k builds for continuous smoke tests
10K builds covers at least 2 weeks of history (if the job is triggered
every 2 mins). Likely whole month, because smoke tests take longer
than 2 mins to run.

Reason why we don't want to keep unlimited history is to save resources
on the concourse VM (disk space, number of records in DB) and prevent
ATC getting stuck/going to 100% CPU as we have seen in 0.74 with
~30K build history.