Specify stemcell version as string
BOSH normalises stemcell versions to strings in v1 manifests, but not
currently for v2 manifests. This causes a misleading prompt about
outstanding changes when you try to use some BOSH operations like:

    /tmp/build/e55deab7 # bosh recreate api/1
    Acting as user 'admin' on deployment 'dcarley' on 'my-bosh'
    You are about to recreate api/1

    Detecting deployment changes
    ----------------------------
    Releases
    No changes

    …

    Stemcells
    No changes

    Cannot perform job management when other deployment changes are present. Please use '--force' to override.

This should be fixed in future bosh_cli versions by cloudfoundry/bosh#1365.

Until then we can, we can ensure that it doesn't happen by specifying the
value as a string. I've added a test to prevent regressions. It probably
doesn't matter if we leave both changes in place after bosh_cli is fixed.