Specify container version as a string.
Otherwise the pipeline fails with "failed to read request: json: cannot
unmarshal number into Go value of type string".

This is because the YAML parser is automagically inferring the type of
the value as a number, which then gets sent in the JSON to the
image_resource container leading to this error.