Remove some dependencies from scripts/Gemfile
`paas-cf/scripts/Gemfile` was forked from `paas-cf/Gemfile` in d271b43
because at the time `bosh_cli` and `cf-uaac` couldn't be loaded at the same
time due to cloudfoundry/cf-uaac#24. This has now been released but changing
it will require more changes to dependencies and testing.

Some of the dependencies that were copied are not needed by `scripts/`, at
either test-time in Travis or run-time in the pipeline. Removing them will
mean less dependencies to manage and a small speed improvement to how fast
the Travis tests and `sync-admin-users` task take to run.