Update concourse browser plugin for new version.
Some of the markup has changed. In earlier versions, there was an
element with an id of 'cli-downloads' that contained the download links
in the lower left corner. With newer versions, this now also displays
the concourse version, and as a result the markup has changed, and
there's no longer an element with the relevant id. Instead, there's a
container with a class of "lower-right-info" that contains all these
details.