Create new Ginkgo suite for concourse/scripts
This is the first step at replacing some (and later all) of our tests for
scripts under `concourse/scripts/` to use Ginkgo and Gomega instead of Rspec
and Aruba+Mimic. We're not terribly happy with the current libraries
because:

Aruba magically hooks into Rspec and is quite complex. The way we're using
it through Rspec (instead of Cucumber, which I'd rather avoid) doesn't seem
to be that commonly used, which is reflected in the unclear API and
documentation. Several people in the team have struggled to write new tests
for scripts because of these issues.

Mimic seems the best choice for creating a real HTTP server with stub
responses that is accessed by a different process in Ruby. However we've run
into problems with it not listening on the correct interface, not being able
to stub 404 bodies, not being able to assert on the contents of requests,
and not being able to query the address it's bound to. We have raised PRs
for some of these issues, but it's unclear whether the project is
maintained.

Alex pointed out that Ginkgo and Gomega have some very good libraries for
executing binaries (gexec) and creating HTTP servers (ghttp). Using them
feels far less "magic" than Aruba and Mimic. I think it's a good fit for us
because we also have other tests in Ginkgo, so we're familiar with the
tooling, and using a separate language from the scripts (which are mostly
shell and Ruby) will help enforce blackbox testing (which is most important
for these scripts where we care about the input/output/exit code).

I'm focussing on converting the scripts that use Mimic at the moment because
the listen bug is preventing us from running them locally under OS X. If
this approach proves successful then we should write all new tests that
would have used Aruba in Ginkgo instead and slowly transition existing tests
over. I have no intention of changing other Rspec tests which don't use
Aruba. So the focus is primarily on `concourse/scripts/`.

Travis only lets you specify a single language for a build. With most of our
other unit tests still being written in Ruby we should keep that as our
primary language for Travis. Go is available in that container, but it's an
older version of 1.4 which might not be supported by Ginkgo in future (they
dropped 1.3). So I'm installing a later version using the `gimme` tool which
Travis use in Go builds. We also need to install Ginkgo and other
dependencies.