Convert bosh_pre_destroy tests to Ginkgo
For the reasons described in the preceding commit which adds a test suite
for the `concourse/scripts/` directory.

The structure and assertions have been copied pretty much like-for-like
which will make it easier to compare whether we are testing the same things.

This script has to be called using `bundle exec` because it uses the
`bosh_cli` gem as a dependency. This makes it a bit slower than the previous
Rspec tests, because they were already within `bundle exec` and didn't have
to initialise it for each test.