Convert kibana_set_utc tests to Ginkgo
For the reasons described in the preceding commit which adds a test suite
for the `concourse/scripts/` directory.

The structure and assertions have been copied pretty much like-for-like
which will make it easier to compare whether we are testing the same things.
Though please note the following:

- I've split the "if <thing>" and "it <thing>" descriptions out to separate
  `Context` and `It` blocks so that I have somewhere to put the
  `BeforeEach`.
- Some of the `Context` blocks could be grouped together at a later date to
  reduce duplication.
- I've added assertions for STDOUT and STDERR content which we didn't have
  before and I believe is important to tell how the script is working.

The script is called directly without `bundle exec` because it doesn't use
any dependencies from rubygems. This is faster because we don't need to
initialise Bundler on each execution. It also prevents later versions of
Bundler from inserting it's own STDOUT content when the script fails:

    ➜  scripts git:(feature/127073275-replace_mimic) ✗ bundle --version
    Bundler version 1.12.5
    ➜  scripts git:(feature/127073275-replace_mimic) ✗ ES_HOST=localhost ES_PORT=9200 bundle exec ./kibana_set_utc.rb
    bundler: failed to load command: ./kibana_set_utc.rb (./kibana_set_utc.rb)
    Errno::ECONNREFUSED: Connection refused - connect(2) for "localhost" port 9200
    …