Increase timeout for bosh_pre_destroy tests
This script sometimes takes more than Eventually's default timeout of 1sec to
return. I think that the need for `bundle exec` is contributing to the total
time, but there's no much we can do about that.

This was first observed by Alex. It could be reproduced by running `ginkgo
--untilItFails`. I tried raising it to 2secs, but that also failed for
Colin. We couldn't get it to fail at 3secs though.

The failure looked like this:

    • Failure [1.006 seconds]
    BoshPreDestroy
    /Users/dcarley/projects/paas/paas-cf/concourse/scripts/bosh_pre_destroy_test.go:94
      two deployments
      /Users/dcarley/projects/paas/paas-cf/concourse/scripts/bosh_pre_destroy_test.go:93
        should return a non-zero exit code and list of deployments [It]
        /Users/dcarley/projects/paas/paas-cf/concourse/scripts/bosh_pre_destroy_test.go:92

        Timed out after 1.000s.
        Expected process to exit.  It did not.

        /Users/dcarley/projects/paas/paas-cf/concourse/scripts/bosh_pre_destroy_test.go:85