Revert change to garden graph_cleanup_threshold_in_mb
This partially reverts commit 81690f6a1143b1ddb6fdbcc886d4b40f0224d837.

This was increased because it was thought that it would be necessary due
to the total size of the containers we use. This turns out not to be the
case because this only applies to container images managed by garden
(which is those using `image:` URLs). Containers managed by baggageclaim
(those using `image_resource:`) are stored separately, and their
expiration is handled separately.