Setup Pingdom email notifications
This relies on the feature added in PR
https://github.com/alphagov/paas-terraform-provider-pingdom/pull/3 on
the terraform pingdom provider.
It allows to add a notification contact to a check.

We add the pingdom contact id of the PaaS team to receive alerts. This
can be improved later by reading this ID from an external source or by
managing the contacts directly with the terraform provider. For the time
being you can extract the ID from the URL on the contact page.

sendnotificationwhendown is set to 2 so we get an alert after 2 check
failures.