Delete local pingdom terraform state
Previously we used to download the state file from the bucket, reupload
it after terraform run, then delete the local copy.

b01f941f16e4d69695d1a8d856b0f20e212619ea changed this to use terraform
built-in S3 storage. Unfortunately it still kept a local copy of the
state file, and this confused terraform when we applied changes to
different environments. We observed for example that resources created
in one environment were updated when applying the changes in another
environment.

This commit deletes the local copy to avoid this issue.