Set check URL for Pingdom HTTPS check
So that Terraform doesn't attempt to set the URL on every run:

    Initialized blank state with remote state enabled!
    Remote state configured and pulled.
    pingdom_check.paas_http_healthcheck: Refreshing state... (ID: 2247550)
    pingdom_check.paas_db_healthcheck: Refreshing state... (ID: 2247551)
    pingdom_check.paas_http_healthcheck: Modifying...
      url: "/" => ""
    pingdom_check.paas_http_healthcheck: Modifications complete

    Apply complete! Resources: 0 added, 1 changed, 0 destroyed.

This happens because the Pingdom API has a default by the Terraform provider
doesn't. Longer term I hope to fix the upstream provider as discussed in
russellcardullo/terraform-provider-pingdom#4, but shorter term this change
will make it easier to see whether Terraform is making real changes.