Pingdom send alerts when back up
This feature is useful to determine whether something needs to be
investigated urgently or how long it was down for.

I believe it was previously configured for all of our checks, by default,
but was disabled either when we changed the contact notification settings in
7024837 or recreated the checks as part of deploying that change.

Managing the attribute explicitly will make sure it is always enabled.