Setup logsearch-for-cloudfoundry filters
We want to be able to parse the CF component logs. For that we
want to use the community filters from logsearch-for-cloudfoundry[1]

We forked that repo of logsearch-for-cloudfoundry to be able to tag it, as it is not tagged upstream.

We used the tag v203.0.0-gds as to correlate it to the last official
version of logsearch-boshrelease, despite we use logsearch-boshrelease v200.0.0.

We add this release to the bosh-cloudfoundry manifest to upload it to bosh.

We configure the job logsearch-for-cloudfoundry-filters to the VMs/jobs:
parser and maintenance.

In parsers we configure the filters in the parser machines by setting
logstash_parser.filters. Note that in logsearch v200 this must be a list
of strings, but more modern versions (> v201)[1] use a hash `key-path`.

In the maintenance job, for elasticsearch, we add the ES type mapping.

[1] https://github.com/cloudfoundry-community/logsearch-for-cloudfoundry/
[2] https://github.com/logsearch/logsearch-boshrelease/commit/e58ebfbafd8e0de075a26f78b0b26d3267f083cc