avail-tests: Test for vegeta lib 404 behaviour
Our availability-tests print the number of non-200 responses and fail if
`metrics.Success` is less than 100%. When looking at the tests and the
documentation for Vegeta, it wasn't clear to me whether `metrics.Success`
included 404 responses or not. This is important because Cloud Foundry's
gorouter will return 404s for an application that exists but has no healthy
instances.

After looking in the code I found that Vegeta does not consider responses
with a status code of 400 or more as successful and that our test is
correct:

- https://github.com/tsenart/vegeta/blob/166e3050f5a8b2f6d3d2cb03e7baa430f019395d/lib/metrics.go#L100-L102

However I think it would be useful to have a test that asserts this
particular behaviour of Vegeta, to save anyone else going on the same hunt
and make us aware in the unlikely event that Vegeta's behaviour changes.