Disable healthcheck DB on DEV
This commit disables provisioning of healthcheck DB on DEV environments to
save space,time and money :). It keeps it enabled on non-dev envs.

Changes consists of default setting in Makefile, that is passed via
pipeline upload script and new condition in deploy-healthcheck to test it.

In an existing development environment the healthcheck-db will persist
until the destroy and autodestroy pipelines will clean it up.