Convert unbind_job tests to Ginkgo
So that we can remove Aruba as described in 7f79755.

The significant differences between the old and new tests are:

- assertions on STDOUT *and* STDERR in each test because they are both
  important when the script is consumed in the pipeline
- re-worded test descriptions because they needed to be split into `Context`
  and `It` so that I had somewhere to assign variables

I did experiment with using a function like `unbind` within each `It` block
that took the job name and input YAML. However I settled on using variables
and a combination of `BeforeEach` and `JustBeforeEach` because I think
having named variables makes the test easier to read.