Convert file_to_yaml tests to Ginkgo
So that we can remove Aruba as described in 7f79755.

The significant difference between the old and new test is that we're no
longer parsing the output YAML and making assertions on individual keys. On
one hand, we're no longer checking that we get valid YAML, but I think this
is outweighed by the benefit of having simpler tests that are easier to
read.